+++ auctions/english-auction/packages/scripts/src/index.ts
@@ -1,4 +1,4 @@
 export * from './helpers/createConfig';
-export * from './helpers/replaceEventOnEnv';
 export * from './helpers/fileNames';
 export * from './types';
+export * from './log';
